requirements.txt
requests
matplotlib
pandas
numpy
langdetect
tensorflow
joblib
scikit-learn
statsmodels
scipy
seaborn
umap-learn
sentence-transformers
ipykernel